Move interface module to SP global utils.
